cambios en estilos de formularios, y vista, ademas cambios en back, de protecistas y pacientes

+++ Odontotal-React/src/pages/pages_admin/dentalProsthetist/ListDentalProsthetist.jsx
@@ -3,18 +3,17 @@ import { ContextGlobal } from '../../../components/utils/global.context'
 import '../../../styles/pagesStyles/ListDentalHygienists.css'
 import Form from './FormToUpdateProsthetist'
 import List from './List'
-import Login from '../../../components/Login'
-import Register from '../../../components/Register'
 import baseUrl from '../../../components/utils/baseUrl.json'
+import '../../../styles/pagesStyles/ListAdmin/ListADentisAdmin.css'
+import Swal from 'sweetalert2';
 
 const ListDentalProsthetist = () => {
-  const { information, user } = useContext(ContextGlobal);
+  const { information, user, jwt } = useContext(ContextGlobal);
 
   const [data, serData] = useState(information);
   const [edition, setedition] = useState(null);
 
   useEffect(() => {
-    // Actualiza el estado de `data` cuando `information` cambia en el contexto global
     serData(information);
   }, [information]);
   
@@ -22,15 +21,59 @@ const ListDentalProsthetist = () => {
     setedition(item);
   };
     
-  const handleEliminar = (item) => {
-    serData((prevState) => prevState.filter((x) => x.id !== item.id));
+  const handleEliminar = async(item) => {
+    // serData((prevState) => prevState.filter((x) => x.id !== item.id));
     const url = baseUrl.url + "/protecistas/" + item.id;
-    const settings = {
-      method: "DELETE",
-    };
-    fetch(url, settings)
-      .then((response) => response.json())
-      .then((error) => console.log(error));
+    const confirmResult = await Swal.fire({
+      title: 'Confirmar datos',
+      text: `¿Esta seguro que desea eliminar los datos del protecista ${item.nombre} ${item.apellido}?`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Aceptar',
+      cancelButtonText: 'Cancelar',
+    });
+    if (confirmResult.isConfirmed) {
+
+         try{
+          const response = await fetch(url, {
+            method: 'DELETE',
+            headers: {
+              'Content-Type': 'application/json',
+              'Authorization': `Bearer ${jwt}`
+            },
+          });
+          
+          if (response.ok) {  
+            Swal.fire(
+              {
+                icon: 'success',
+                title: 'Datos eliminado correctamente',
+                showCancelButton: false,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Aceptar',
+              }
+            ).then((result) => {
+              if (result.isConfirmed) {
+                window.location.pathname="/ListaDeProtecistas"
+              }
+            })
+          } else {
+            console.error('Error al enviar los datos');
+            Swal.fire({
+              icon: "error",
+              title: "Error al eliminar",
+
+            });
+            }
+        }
+        catch (error) {
+          console.error('Error en la conexión', error);
+          }
+            }
+            window.location.pathname="/ListaDeProtecistas"
   };
 
   const handleGuardar = (item) => {
@@ -49,11 +92,11 @@ const ListDentalProsthetist = () => {
   }
 
   return (
-    <div style={{display: "flex", flexDirection: "column"}}>
+    <div className='main'>
       {user?.rol === "ADMIN" &&
       <>
       {edition ? (
-        <Form data={edition} onGuardar={handleGuardar} informacionCompleta={data}  onCancelar={handleCancelar} />
+        <Form data={edition} onGuardar={handleGuardar} informacionCompleta={data}  onCancelar={handleCancelar} jwt={jwt}/>
       ) : (
         <List data={data} onEditar={handleEditar} onEliminar={handleEliminar} />
       )}
