+++ Odontotal-React/src/pages/pages_admin/dentist/ListDentistAdmin.jsx
@@ -1,11 +1,10 @@
 import React, { useContext, useEffect, useState } from 'react'
 import { ContextGlobal } from '../../../components/utils/global.context'
-// import '../../../styles/pagesStyles/ListDentalHygienists.css'
+import '../../../styles/pagesStyles/ListAdmin/ListADentisAdmin.css'
 import Form from './FormToUpdateDentist'
 import List from './List'
-// import Login from '../../../components/Login'
-// import Register from '../../../components/Register'
 import baseUrl from '../../../components/utils/baseUrl.json'
+import Swal from 'sweetalert2';
 
 const ListDentalProsthetist = () => {
   const { information, user, jwt} = useContext(ContextGlobal);
@@ -21,19 +20,59 @@ const ListDentalProsthetist = () => {
     setedition(item);
   };
     
-  const handleEliminar = (item) => {
-    serData((prevState) => prevState.filter((x) => x.id !== item.id));
-    const url = baseUrl.url + "/odontologos/" + item.id;
+  const handleEliminar = async(item) => {
     
-    const settings = {
-      method: "DELETE",
-      headers: {
-        'Authorization': 'Bearer ' + jwt
-      },
-    };
-    fetch(url, settings)
-      .then((response) => response.json())
-      .then((error) => console.log(error));
+    const url = baseUrl.url + "/odontologos/" + item.id;
+    const confirmResult = await Swal.fire({
+      title: 'Confirmar datos',
+      text: `¿Esta seguro que desea eliminar los datos del protecista ${item.nombre} ${item.apellido}?`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Aceptar',
+      cancelButtonText: 'Cancelar',
+    });
+    if (confirmResult.isConfirmed) {
+
+         try{
+          const response = await fetch(url, {
+            method: 'DELETE',
+            headers: {
+              'Content-Type': 'application/json',
+              'Authorization': `Bearer ${jwt}`
+            },
+          });
+          
+          if (response.ok) {  
+            Swal.fire(
+              {
+                icon: 'success',
+                title: 'Datos eliminado correctamente',
+                showCancelButton: false,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Aceptar',
+              }
+            ).then((result) => {
+              if (result.isConfirmed) {
+                window.location.pathname="/ListaDeOdontologos"
+              }
+            })
+          } else {
+            console.error('Error al enviar los datos');
+            Swal.fire({
+              icon: "error",
+              title: "Error al eliminar",
+
+            });
+            }
+        }
+        catch (error) {
+          console.error('Error en la conexión', error);
+          }
+            }
+            window.location.pathname="/ListaDeOdontologos"
   };
 
   const handleGuardar = (item) => {
@@ -52,7 +91,7 @@ const ListDentalProsthetist = () => {
   }
 
   return (
-    <div className='list-odonto-container' style={{display: "flex", flexDirection: "column"}}>
+    <div className='main'>
       {user?.rol === "ADMIN" &&
       <>
       {edition ? (
